+++ tools/editor/run.ts
@@ -1,0 +1,6 @@
+import * as sh from "shelljs";
+
+sh.exec(`
+    GOOS=linux go build -o editor
+    GOOS=windows go build -o editor.exe
+`);